fix(aws): simplify region filtering follow-up
