+++ cartography/intel/aws/sagemaker/__init__.py
@@ -18,7 +18,7 @@ from cartography.intel.aws.sagemaker.notebook_instances import sync_notebook_ins
 from cartography.intel.aws.sagemaker.training_jobs import sync_training_jobs
 from cartography.intel.aws.sagemaker.transform_jobs import sync_transform_jobs
 from cartography.intel.aws.sagemaker.user_profiles import sync_user_profiles
-from cartography.intel.aws.util.common import (
+from cartography.intel.aws.util.service_regions import (
     filter_regions_to_supported_service_regions,
 )
 from cartography.util import timeit
@@ -57,12 +57,7 @@ def sync(
             regions,
         )
     )
-    if sagemaker_regions == regions and not unsupported_regions and regions:
-        logger.warning(
-            "Could not determine available SageMaker regions for account '%s'. Continuing with requested regions.",
-            current_aws_account_id,
-        )
-    elif unsupported_regions:
+    if unsupported_regions:
         logger.info(
             "Skipping SageMaker sync for account '%s' in unsupported regions: %s",
             current_aws_account_id,
